Moved message icons around
Got rid of attachments icon on the chat header. Removed the camera icon on the chat footer and replace it with the attachments icon.
One slight bug happening however. When first loading a chat room, it starts the messages from oldest first, and have to scroll down to the button to see the newest messages. Should load into the chat with the newest messages first.

Co-Authored-By: karimsaraipour <75566475+karimsaraipour@users.noreply.github.com>